// models/repoFile.ts
export default interface RepoFile {
  name: string;
  path: string;
  sha: string;
  size: number;
  url: string;
  html_url: string;
  git_url: string;
  download_url: string;
  type: "file";
  content: string;
  encoding: string;
  _links: {
    self: string;
    git: string;
    html: string;
  };
}